+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Changed
 
 - Accordion list component accepts itemKey property, and uses this to generate item keys. By setting the `itemKey`, more than one accordion list on the page can be used.
+- Add shadow to modal (dialog) border.
+- Cloze question can include line breaks with role `lineBreak`.
+
 
 ## 1.6.2 - 2021-11-22